+++ src/1_Page/Profile/ui/AwardDashBoard/type.d.ts
