+++ src/1_Page/Profile/ui/PlayerDashBoard/model/type.d.ts
@@ -1,0 +1,5 @@
+export type UseModifyHandlerProps = {
+  userInfoForm: UserInfoForm;
+  reset: UseFormReset<UserInfoForm>;
+  inputBackupDataRef: RefObject<UserInfoForm>;
+};
